requirements.txt
crewai>=0.28.8
crewai-tools>=0.1.6
selenium>=4.15.2
beautifulsoup4>=4.12.2
requests>=2.31.0
pandas>=2.1.4
python-dotenv>=1.0.0
langchain>=0.1.10
langchain-openai>=0.0.5
langchain-community>=0.0.10
streamlit>=1.29.0
plotly>=5.17.0
openpyxl>=3.1.2
googlemaps>=4.10.0 